pyproject.toml
[project]
name = "multimidia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "ffmpeg-python>=0.2.0",
    "huggingface-hub>=0.36.0",
    "lightning==2.5.5",
    "openai-whisper>=20250625",
    "pyannote-audio==4.0.1",
    "python-multipart>=0.0.20",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchcodec>=0.8.0",
    "torchmetrics>=1.0.3",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchcodec = { index = "torchcodec" }
torchmetrics = { index = "torchmetrics" }
tqdm = { index = "tqdm" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/rocm6.4"

[[tool.uv.index]]
name = "torchcodec"
url = "https://download.pytorch.org/whl/cpu/"

[[tool.uv.index]]
name = "torchmetrics"
url = "https://pypi.org/simple/"

[[tool.uv.index]]
name = "tqdm"
url = "https://pypi.org/simple/"

[tool.uv.pip]
index-strategy = "unsafe-best-match"
